Rename Eoa -> ExternallyOwnedAccount
