+++ ethdk/src/Networks/index.ts
@@ -1,5 +1,5 @@
 import { localhost as blsLocalhost } from '../Bls/BlsNetworks'
-import { localhost as eoaLocalhost } from '../Eoa/EoaNetworks'
+import { localhost as eoaLocalhost } from '../ExternallyOwnedAccount/ExternallyOwnedAccountNetworks'
 
 export const BLS_NETWORKS = {
   localhost: blsLocalhost,
